% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_palette.R
\name{create_palette}
\alias{create_palette}
\title{Create a palette of colours for a graph}
\usage{
create_palette(colour)
}
\arguments{
\item{colour}{A vector containing the graph objects which require different colours}
}
\value{
A vector with HEX code colours for a graph.
The colour palette generated is created by PHS to be accessible when used in the order provided.
}
\description{
Create a palette of colours for a graph
}
\examples{
create_palette(c("2021/22", "2022/23", "2023/24"))
# [1] "#12436D" "#28A197" "#801650"
}
